Factor out widgets from frontend/src/main.rs
